requirements.txt
streamlit
pymongo 
transformers 
langchain 
langchain_core 
google-generativeai 
langchain-google-genai 
python-dotenv